348 Split ABM tests into a more structured way (#457)
* Split ABM tests into a more structured way
